pyproject.toml
[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[project]
name = "llamaparse-minimax"
version = "0.1.0"
description = "LlamaParse + MiniMax pipeline for PDF parsing and plain-English financial document summarization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "llama-cloud-services>=0.6.89",
    "llama-index-workflows>=0.5.0",
    "pandas>=2.3.3",
    "tabulate>=0.9.0",
]

[tool.uv.build-backend]
module-name = "llamaparse_gemini"

[project.scripts]
run-workflow = "llamaparse_gemini.main:main"
